# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  dtos.CredenciaisDto:
    properties:
      email:
        type: string
      senha:
        type: string
    type: object
  dtos.PassageiroDto:
    properties:
      cpf:
        type: string
      email:
        type: string
      nome:
        type: string
      senha:
        type: string
    type: object
  models.Passageiro:
    properties:
      cpf:
        type: string
      email:
        type: string
      idPassageiro:
        type: integer
      nome:
        type: string
      senha:
        type: string
    type: object
  models.Passagem:
    properties:
      dataCompra:
        type: string
      idPassageiro:
        type: integer
      idPassagem:
        type: integer
      idVoo:
        type: integer
      numeroAcento:
        type: integer
    type: object
  models.ResponseToken:
    properties:
      token:
        type: string
    type: object
  models.Voo:
    properties:
      capacidade:
        type: integer
      destino:
        type: string
      horaSaida:
        type: string
      idVoo:
        type: integer
    type: object
host: '{{.Host}}'
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: verifica na base se existe as credenciais passadas, e retorna um
        token caso exista.
      parameters:
      - description: Login
        in: body
        name: Passageiro
        required: true
        schema:
          $ref: '#/definitions/dtos.CredenciaisDto'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseToken'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: realizar login
      tags:
      - Auth
  /passageiro:
    get:
      consumes:
      - application/json
      description: Retorna todos os passageiros cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Passageiro'
            type: array
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a passageiro
      tags:
      - Passageiro
    post:
      consumes:
      - application/json
      description: Metodo que cadastrar passageiro
      parameters:
      - description: Passageiro
        in: body
        name: Passageiro
        required: true
        schema:
          $ref: '#/definitions/dtos.PassageiroDto'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Passageiro cadastrado
          schema:
            type: string
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a passageiro
      tags:
      - Passageiro
    put:
      consumes:
      - application/json
      description: Metodo para atualizar um voo
      parameters:
      - description: Passageiro
        in: body
        name: Passageiro
        required: true
        schema:
          $ref: '#/definitions/models.Passageiro'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Passageiro atualizado
          schema:
            type: string
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a voo
      tags:
      - Passageiro
  /passageiro/{idPassageiro}:
    get:
      consumes:
      - application/json
      description: Retorna o passageiros cadastrado com base no idPassageiro passado
      parameters:
      - description: IdPassageiro
        in: path
        name: idVoo
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Passageiro'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a passageiro
      tags:
      - Passageiro
  /passagem:
    delete:
      consumes:
      - application/json
      description: Metodo para excluir uma pasagem
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Passagem'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a passagens
      tags:
      - Passagem
    get:
      consumes:
      - application/json
      description: Metodo para buscar passagem passando o id da passagem como referencia
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Passagem'
            type: array
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a passagens
      tags:
      - Passagem
    post:
      consumes:
      - application/json
      description: Metodo para registrar uma pasagem
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Passagem'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a passagens
      tags:
      - Passagem
    put:
      consumes:
      - application/json
      description: Metodo atualiza uma passagem
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Passagem'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a passagens
      tags:
      - Passagem
  /voo:
    delete:
      consumes:
      - application/json
      description: Metodo que cadastrar voos
      parameters:
      - description: IdVoo
        in: path
        name: idVoo
        required: true
        type: integer
      - description: IdVoo
        in: path
        name: destino
        required: true
        type: string
      - description: IdVoo
        in: path
        name: horaSaida
        required: true
        type: string
      - description: IdVoo
        in: path
        name: capacidade
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Voo'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a voo
      tags:
      - voo
    get:
      consumes:
      - application/json
      description: Retorna todos os voos cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Voo'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a voo
      tags:
      - voo
    post:
      consumes:
      - application/json
      description: Metodo que cadastrar voos
      parameters:
      - description: IdVoo
        in: path
        name: idVoo
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Voo'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a voo
      tags:
      - voo
    put:
      consumes:
      - application/json
      description: Metodo para atualizar um voo
      produces:
      - application/json
      responses:
        "200": {}
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a voo
      tags:
      - voo
  /voo/{idVoo}:
    get:
      consumes:
      - application/json
      description: Retorna o voo cadastrado com base no idVoo passado
      parameters:
      - description: IdVoo
        in: path
        name: idVoo
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Voo'
            type: object
        "400":
          description: Requisição invalida
          schema:
            type: string
        "401":
          description: Não autorizado
          schema:
            type: string
        "404":
          description: Nenhum registro encontrado.
          schema:
            type: string
      summary: Show a voo
      tags:
      - voo
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"